requirements.txt
psutil==5.9.3
mutagen==1.46.0
py-cpuinfo==9.0.0
hurry.filesize==0.9